move onboarding-ms to connector
