+++ connector/rest/src/main/java/it/pagopa/selfcare/onboarding/connector/PartyConnectorImpl.java
@@ -1,7 +1,6 @@
 package it.pagopa.selfcare.onboarding.connector;
 
 import it.pagopa.selfcare.commons.base.logging.LogUtils;
-import it.pagopa.selfcare.commons.base.utils.InstitutionType;
 import it.pagopa.selfcare.onboarding.connector.api.PartyConnector;
 import it.pagopa.selfcare.onboarding.connector.model.RelationshipInfo;
 import it.pagopa.selfcare.onboarding.connector.model.RelationshipsResponse;
@@ -11,10 +10,8 @@ import it.pagopa.selfcare.onboarding.connector.model.institutions.OnboardingReso
 import it.pagopa.selfcare.onboarding.connector.model.onboarding.*;
 import it.pagopa.selfcare.onboarding.connector.rest.client.MsCoreOnboardingApiClient;
 import it.pagopa.selfcare.onboarding.connector.rest.client.MsCoreTokenApiClient;
-import it.pagopa.selfcare.onboarding.connector.rest.client.MsOnboardingApiClient;
 import it.pagopa.selfcare.onboarding.connector.rest.client.PartyProcessRestClient;
 import it.pagopa.selfcare.onboarding.connector.rest.mapper.InstitutionMapper;
-import it.pagopa.selfcare.onboarding.connector.rest.mapper.OnboardingMapper;
 import it.pagopa.selfcare.onboarding.connector.rest.model.InstitutionUpdate;
 import it.pagopa.selfcare.onboarding.connector.rest.model.*;
 import lombok.extern.slf4j.Slf4j;
@@ -47,10 +44,6 @@ class PartyConnectorImpl implements PartyConnector {
     private final PartyProcessRestClient restClient;
     private final InstitutionMapper institutionMapper;
 
-    private final MsOnboardingApiClient msOnboardingApiClient;
-
-    private final OnboardingMapper onboardingMapper;
-
     private static final BinaryOperator<InstitutionInfo> MERGE_FUNCTION =
             (inst1, inst2) -> inst1.getUserRole().compareTo(inst2.getUserRole()) < 0 ? inst1 : inst2;
     private static final Function<OnboardingResponseData, InstitutionInfo> ONBOARDING_DATA_TO_INSTITUTION_INFO_FUNCTION = onboardingData -> {
@@ -80,25 +73,11 @@ class PartyConnectorImpl implements PartyConnector {
     };
 
     @Autowired
-    public PartyConnectorImpl(MsCoreTokenApiClient msCoreTokenApiClient, MsCoreOnboardingApiClient msCoreOnboardingApiClient, PartyProcessRestClient restClient, InstitutionMapper institutionMapper, MsOnboardingApiClient msOnboardingApiClient, OnboardingMapper onboardingMapper) {
+    public PartyConnectorImpl(MsCoreTokenApiClient msCoreTokenApiClient, MsCoreOnboardingApiClient msCoreOnboardingApiClient, PartyProcessRestClient restClient, InstitutionMapper institutionMapper) {
         this.msCoreTokenApiClient = msCoreTokenApiClient;
         this.msCoreOnboardingApiClient = msCoreOnboardingApiClient;
         this.restClient = restClient;
         this.institutionMapper = institutionMapper;
-        this.msOnboardingApiClient = msOnboardingApiClient;
-        this.onboardingMapper = onboardingMapper;
-    }
-
-
-    @Override
-    public void onboarding(OnboardingData onboardingData) {
-        if (onboardingData.getInstitutionType() == InstitutionType.PA) {
-            msOnboardingApiClient._onboardingPaPost(onboardingMapper.toOnboardingPaRequest(onboardingData));
-        } else if (onboardingData.getInstitutionType() == InstitutionType.PSP) {
-            msOnboardingApiClient._onboardingPspPost(onboardingMapper.toOnboardingPspRequest(onboardingData));
-        } else {
-            msOnboardingApiClient._onboardingPost(onboardingMapper.toOnboardingDefaultRequest(onboardingData));
-        }
     }
 
     @Override
