+++ web/src/main/java/it/pagopa/selfcare/onboarding/web/controller/TokenController.java
@@ -6,6 +6,7 @@ import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
 import it.pagopa.selfcare.commons.base.logging.LogUtils;
 import it.pagopa.selfcare.onboarding.core.TokenService;
+import it.pagopa.selfcare.onboarding.web.model.OnboardingMode;
 import it.pagopa.selfcare.onboarding.web.model.TokenVerifyResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -13,6 +14,8 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+
+import java.util.Objects;
 
 @Slf4j
 @RestController
@@ -62,10 +65,13 @@ public class TokenController {
     @PostMapping(value = "/{tokenId}/complete")
     public ResponseEntity<Void> complete(@ApiParam("${swagger.tokens.tokenId}")
                                                    @PathVariable(value = "tokenId") String tokenId,
-                                                   @RequestPart MultipartFile contract) {
+                                                   @RequestPart MultipartFile contract,
+                                                    @RequestParam(value = "mode", required = false) OnboardingMode mode) {
         log.trace("complete Token start");
         log.debug(LogUtils.CONFIDENTIAL_MARKER, "complete Token tokenId = {}, contract = {}", tokenId, contract);
-        tokenService.completeToken(tokenId, contract);
+        if(Objects.isNull(mode) || mode.equals(OnboardingMode.SYNC))
+            tokenService.completeToken(tokenId, contract);
+        else tokenService.completeTokenAsync(tokenId, contract);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
